Switch from uint8_t to uint32_t

+++ audio_rec.c
@@ -17,7 +17,7 @@
 #define DMA_CHANNEL_ADC 0
 
 typedef struct {
-	uint8_t *pcm_buf;
+	uint32_t *pcm_buf;
 	unsigned int pcm_bufsize;
 } Dma;
 
@@ -30,8 +30,8 @@ DMA_CB_TypeDef cb;
 volatile bool transferActive;	// TODO: why is this volatile?
 bool enable_transfer = false;
 
-uint8_t *end_of_data;
-uint8_t *read_pointer;	// a pointer to the chunk which should be read next
+uint32_t *end_of_data;
+uint32_t *read_pointer;	// a pointer to the chunk which should be read next
 
 #define SAMPLE_RATE 8000  // 8000 hz sample rate
 
@@ -62,7 +62,7 @@ void transferComplete(unsigned int channel, bool primary, void *user)
 {
 	Dma *dma = (Dma *) user;
 
-	uint8_t *cyclic_buf = dma->pcm_buf;
+	uint32_t *cyclic_buf = dma->pcm_buf;
 
 	static int p = 2 * NUMOF_ADC_SAMPLES;
 
@@ -219,7 +219,7 @@ void setupDma(Dma *dma)
 	DMA_CfgDescr(0, true, &descrCfg);
 	DMA_CfgDescr(0, false, &descrCfg);
 
-	uint8_t *cyclic_buf = dma->pcm_buf; // temporary work-around
+	uint32_t *cyclic_buf = dma->pcm_buf; // temporary work-around
 	
 	
 	/* Enabling PingPong Transfer*/
@@ -390,7 +390,7 @@ void setupOpAmp(void)
  * using ping-pong transfer.
  *****************************************************************************/
 // TODO: rewrite this using start_recording and stop_recording....
-void record(uint8_t *pcm_buf, unsigned int pcm_bufsize, unsigned int numof_secs)
+void record(uint32_t *pcm_buf, unsigned int pcm_bufsize, unsigned int numof_secs)
 {
 	// the number of ping pong transfers that will occur in numof_secs time
 	int transfer_limit = (SAMPLE_RATE / NUMOF_ADC_SAMPLES) * numof_secs;
@@ -406,7 +406,7 @@ void record(uint8_t *pcm_buf, unsigned int pcm_bufsize, unsigned int numof_secs)
 }
 
 // TODO: test this actually works...
-void start_recording(uint8_t *pcm_buf, unsigned int pcm_bufsize)
+void start_recording(uint32_t *pcm_buf, unsigned int pcm_bufsize)
 {
 	enable_transfer = true;
 
@@ -434,8 +434,7 @@ void stop_recording(void)
 	DMA_Reset();	// clean up after DMA transfers
 }
 
-// TODO: implement this...
-bool read_chunk(uint8_t **chunk)
+bool read_chunk(uint32_t **chunk)
 {
 	if (read_pointer != end_of_data) {
 		*chunk = read_pointer;
